% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{check_eaf}
\alias{check_eaf}
\title{check_eaf}
\usage{
check_eaf(ssp)
}
\arguments{
\item{ssp:}{a list of all study files which contain summary statistics data in different ancestries}
}
\value{
the revised list of all study files
}
\description{
Check whether all efficient groups share the same efficient allele
}
