% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceMatrix.R
\name{dist_pop}
\alias{dist_pop}
\title{dist_pop}
\usage{
dist_pop(X)
}
\arguments{
\item{X:}{a matrix (N * the number of study files) of which entry is EAF}
}
\value{
a matrix named double centering for multi-dimension scaling
}
\description{
Calculating the multi dimensional scaling
}
