% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceMatrix.R
\name{double_center}
\alias{double_center}
\title{double_center}
\usage{
double_center(D)
}
\arguments{
\item{D:}{matrix of multi-dimension scaling}
}
\value{
a matrix through double centering
}
\description{
double centering for multi-dimension scaling
}
