% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_mega_Rfunction.R
\name{getN_AverageEAF}
\alias{getN_AverageEAF}
\title{getN_AverageEAF}
\usage{
getN_AverageEAF(ssp, ssf)
}
\arguments{
\item{ssp:}{the revised list of all study files}

\item{ssf:}{the filtered variants names}
}
\value{
a list which contains the sum of sample size among all studies and the average of eaf
}
\description{
summarize the sample size and eaf
}
