% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_mega_Rfunction.R
\name{lr_w}
\alias{lr_w}
\title{linear regression}
\usage{
lr_w(Y, X, W)
}
\arguments{
\item{Y:}{BETA}

\item{X:}{pcs principal components}

\item{W:}{SE standard errors}
}
\value{
a list of regression result for one variant
}
\description{
linear regression for each variants
}
