% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_mega_Rfunction.R
\name{MR_mega}
\alias{MR_mega}
\title{MR_mega}
\usage{
MR_mega(
  data_name,
  data_loc = NULL,
  qt = TRUE,
  pcCount = 1,
  nCores,
  parallel = TRUE
)
}
\arguments{
\item{data_name:}{an input file which contains the list of all study files}

\item{data_loc:}{a character vector of working directory}

\item{qt:}{an indicator of whether the trait is quatitative or not}

\item{pcCount:}{the number of principal components}

\item{nCores:}{the number of cores which are used for running parallel}

\item{parallel:}{an indicator of whether proceeed parallel}
}
\value{
a csv file which contains statistics and p values under different hyporthesis testings
}
\description{
Meta regression
}
