% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{readInputfile}
\alias{readInputfile}
\title{readInputfile}
\usage{
readInputfile(data_name, qt = TRUE, data_loc = NULL)
}
\arguments{
\item{data_name:}{an input file which contains the list of all study files}

\item{qt:}{an indicator of whether the trait is quatitative or dichotomous}

\item{data_loc:}{a character vector of working directory}
}
\value{
a list of all study files which contain summary statistics data in different ancestries
}
\description{
read file which contains names of different ancestry groups summary statistics
}
