% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceMatrix.R
\name{sel_MafChrPos}
\alias{sel_MafChrPos}
\title{sel_MafChrPos}
\usage{
sel_MafChrPos(ssp, sp)
}
\arguments{
\item{ssp:}{a list of all study files which contain summary statistics data in different ancestries}

\item{sp:}{a vector which contains all variants names}
}
\value{
a filtered list of all study files
}
\description{
filt out some variants of which summary statistics are not included
}
